Add a class to report on unhandled exceptions for KCL threads
Added ThreadExceptionReporter which will attempt to report on any
unhandled exceptions in threads the reporter is installed for.  It's
important to remember that this class only reports on uncaught
exceptions.  The JVM will still kill the thread that recieved the exception.

Updated all thread factories to use instances of a
ThreadExceptionReporter to report any uncaught exceptions.