Fix `logOnErrorReadTimeoutWarning` formatting
The formatting of the `logOnErrorReadTimeoutWarning` was inconsistent in where it had spaces and where it didn't. This caused words to be smashed together in the log message.

Example output:
```
onError().  Cancelling subscription, and marking self as failed. KCL will recreate the subscription as neccessary to continue processing. If you are seeing this warning frequently consider increasing the SDK timeouts by providing an OverrideConfiguration to the kinesis client. Alternatively youcan configure LifecycleConfig.readTimeoutsToIgnoreBeforeWarning to suppressintermittent ReadTimeout warnings. Last successful request details -- request id - UNKNOWN, timestamp - 2021-12-14T17:41:00.169Z
```
specifically note the `youcan` and `suppressintermittent`.

This change corrects the formatting of this log output to not include these smashed together words.

It also corrects a type in the word `necessary` (was `neccessary`).